add assert

+++ src/kernel/include/onix/kernel/interrupt.h
@@ -144,6 +144,6 @@ extern InterruptGate idt[IDT_SIZE];
 extern Pointer idt_ptr;
 extern InterruptHandler interrupt_entry_table[IDT_SIZE];
 extern InterruptHandler handler_table[IDT_SIZE];
-extern u64 __clock_counter;
+extern u32 __clock_counter;
 
 #endif
